Goal to receive an MKS style POST /upload payload and convert it into legacy style M28/M29 command sequences.
This receives payloads from a pure socket server [intended to run on
port 80 because prusa slicer expects MKS web interface to be on port 80]

it looks at the first line and determines if it is an MKS style POST
upload.  If not, it will respond with an http redirect to point to the
actual web server (on my system port 8000).

If it is an MKS style POST future work will try to parse and load to
M28/M29 commands