dont allow new upload packets in while still processing the old one
start pulling in some of the synchronous serial handling code from the
M28 codepath.

NOTE:  at this point I determined that upload file payloads were not
being processed via the serial port and decided to switch gears and try
to emulate the ESP01 interface instead.

Will start a new branch for that and pull a lot of this code over as
needed.