change timeout for chitu to respond
add U102 and U103 reponses

Note, Chitu HB seems to successfully send files with correct checksums now even handling
retransmissions if requested or if a packet gets lost.  Makes sense
since it is a UDP protocol.

Note there is a race between Chitu HB "info" recurring requests and
Chitu HB tasks like file listings.  They must be getting interleaved.

Need to add some logic to prevent requests if are in the midst of a
current request.  This is needed anyway to support MKS upload mode and
not to have conflict with the esp3d ui.